Ignore routes marked `administrate: false`
Problem:

Users often define routes in the `admin` namespace
that they don't want to show up in the sidebar.
At the moment, Administrate creates a sidebar link
for all routes in the `admin` namepsace,
without a way to ignore them.

Solution:

Add an `administrate: false` option for dashboard routes
that will prevent Administrate
from creating a sidebar link for the resource.